Refactor: Modularity for HomeActivity and MainActivity
